+++ CHANGELOG.md
@@ -1,15 +1,28 @@
 # Changelog
 
+This file contains the changes for the projects.
 
-## 0.2.1
+- [CoLib.Common](#colibcommon)
+- [CoLib.Common.EF](#colibcommonef)
+
+## CoLib.Common
+
+### 0.2.1
 
 - Description for NuGet-Package updated
 
-## 0.2.0
+### 0.2.0
 
 - NEW: Serializable Dictionary added (#1)
 
-## 0.1.0
+### 0.1.0
 
 - First release
 - NEW: StringExtension.TrimOrEmpty()
+
+
+## CoLib.Common.EF
+
+### 1.0.0
+
+- NEW: DbModelBuilderExtension to use TablePrefix